+++ src/saivenky/neural/IInputLayer.java
@@ -4,5 +4,5 @@ package saivenky.neural;
  * Created by saivenky on 2/21/17.
  */
 public interface IInputLayer extends ILayer {
-    void setInput(double[][] input);
+    void setInput(float[][] input);
 }
