+++ src/saivenky/neural/ILayer.java
@@ -6,6 +6,6 @@ package saivenky.neural;
 public interface ILayer {
     void feedforward();
     void backpropagate();
-    void gradientDescent(double rate);
+    void gradientDescent(float rate);
 }
