+++ src/saivenky/neural/INeuralNetwork.java
@@ -4,8 +4,8 @@ package saivenky.neural;
  * Created by saivenky on 3/4/17.
  */
 public interface INeuralNetwork {
-    double[][] getPredicted();
-    void run(double[][] input);
-    void update(double rate);
-    void train(double[][] input, double[][] output);
+    float[][] getPredicted();
+    void run(float[][] input);
+    void update(float rate);
+    void train(float[][] input, float[][] output);
 }
